requirements.txt
Flask
PyTorch
torchvision
Pillow